all semantic actions out other than one in var_decls_grammar
