+++ src/stan/lang/grammars/program_grammar.hpp
@@ -1,21 +1,20 @@
 #ifndef STAN_LANG_GRAMMARS_PROGRAM_GRAMMAR_HPP
 #define STAN_LANG_GRAMMARS_PROGRAM_GRAMMAR_HPP
 
-#include <boost/spirit/include/qi.hpp>
-
 #include <stan/lang/ast.hpp>
 #include <stan/lang/grammars/whitespace_grammar.hpp>
 #include <stan/lang/grammars/expression_grammar.hpp>
 #include <stan/lang/grammars/var_decls_grammar.hpp>
 #include <stan/lang/grammars/statement_grammar.hpp>
 #include <stan/lang/grammars/functions_grammar.hpp>
+#include <boost/spirit/include/qi.hpp>
 
 #include <string>
 #include <sstream>
-#include <utility>
 #include <vector>
 
 namespace stan {
+
   namespace lang {
 
     template <typename Iterator>
@@ -23,20 +22,15 @@ namespace stan {
       : boost::spirit::qi::grammar<Iterator,
                                    program(),
                                    whitespace_grammar<Iterator> > {
-      program_grammar(const std::string& model_name);
-
-      // global info for parses
       std::string model_name_;
       variable_map var_map_;
       std::stringstream error_msgs_;
-
-      // grammars
       expression_grammar<Iterator> expression_g;
       var_decls_grammar<Iterator> var_decls_g;
       statement_grammar<Iterator> statement_g;
       functions_grammar<Iterator> functions_g;
 
-      // rules
+      program_grammar(const std::string& model_name);
 
       boost::spirit::qi::rule<Iterator,
                               std::vector<var_decl>(),
