+++ src/stan/lang/grammars/statement_2_grammar_def.hpp
@@ -1,83 +1,17 @@
 #ifndef STAN_LANG_GRAMMARS_STATEMENT_2_GRAMMAR_DEF_HPP
 #define STAN_LANG_GRAMMARS_STATEMENT_2_GRAMMAR_DEF_HPP
 
-#include <boost/lexical_cast.hpp>
-#include <boost/config/warning_disable.hpp>
-#include <boost/fusion/include/adapt_struct.hpp>
-#include <boost/fusion/include/std_pair.hpp>
-#include <boost/spirit/include/classic_position_iterator.hpp>
-#include <boost/spirit/include/phoenix_core.hpp>
-#include <boost/spirit/include/phoenix_function.hpp>
-#include <boost/spirit/include/phoenix_fusion.hpp>
-#include <boost/spirit/include/phoenix_object.hpp>
-#include <boost/spirit/include/phoenix_operator.hpp>
-#include <boost/spirit/include/phoenix_stl.hpp>
-#include <boost/spirit/include/qi.hpp>
-#include <boost/spirit/include/qi_numeric.hpp>
-#include <boost/spirit/include/support_multi_pass.hpp>
-#include <boost/tuple/tuple.hpp>
-#include <boost/variant/apply_visitor.hpp>
-#include <boost/variant/recursive_variant.hpp>
-
 #include <stan/lang/ast.hpp>
 #include <stan/lang/grammars/common_adaptors_def.hpp>
-#include <stan/lang/grammars/expression_grammar.hpp>
 #include <stan/lang/grammars/statement_grammar.hpp>
 #include <stan/lang/grammars/statement_2_grammar.hpp>
-#include <stan/lang/grammars/whitespace_grammar.hpp>
-
-#include <cstddef>
-#include <iomanip>
-#include <iostream>
-#include <istream>
-#include <map>
-#include <set>
+#include <boost/spirit/include/qi.hpp>
+#include <boost/spirit/include/phoenix_core.hpp>
 #include <sstream>
-#include <stdexcept>
-#include <string>
-#include <utility>
-#include <vector>
 
 namespace stan {
 
   namespace lang {
-
-    struct add_conditional_condition {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3, typename T4>
-      struct result<F(T1, T2, T3, T4)> { typedef void type; };
-      void operator()(conditional_statement& cs,
-                      const expression& e,
-                      bool& pass,
-                      std::stringstream& error_msgs) const {
-        if (!e.expression_type().is_primitive()) {
-          error_msgs << "conditions in if-else statement must be"
-                     << " primitive int or real;"
-                     << " found type=" << e.expression_type()
-                     << std::endl;
-          pass = false;
-          return;
-        }
-        cs.conditions_.push_back(e);
-        pass = true;
-        return;
-      }
-    };
-    boost::phoenix::function<add_conditional_condition>
-    add_conditional_condition_f;
-
-    struct add_conditional_body {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2>
-      struct result<F(T1, T2)> { typedef void type; };
-      void operator()(conditional_statement& cs,
-                      const statement& s) const {
-        cs.bodies_.push_back(s);
-      }
-    };
-    boost::phoenix::function<add_conditional_body> add_conditional_body_f;
-
-
 
     template <typename Iterator>
     statement_2_grammar<Iterator>::statement_2_grammar(variable_map& var_map,
@@ -102,9 +36,7 @@ namespace stan {
       // _r2 source of variables allowed for assignments
       // set to true if sample_r are allowed
       statement_2_r.name("statement");
-      statement_2_r
-        %= conditional_statement_r(_r1, _r2, _r3);
-
+      statement_2_r %= conditional_statement_r(_r1, _r2, _r3);
 
       conditional_statement_r.name("if-else statement");
       conditional_statement_r
