+++ src/stan/lang/grammars/var_decls_grammar_def.hpp
@@ -1,24 +1,11 @@
 #ifndef STAN_LANG_GRAMMARS_VAR_DECLS_GRAMMAR_DEF_HPP
 #define STAN_LANG_GRAMMARS_VAR_DECLS_GRAMMAR_DEF_HPP
 
-#include <boost/spirit/include/qi.hpp>
-// FIXME: get rid of unused include
-#include <boost/spirit/include/phoenix_core.hpp>
-#include <boost/spirit/include/phoenix_function.hpp>
-#include <boost/spirit/include/phoenix_fusion.hpp>
-#include <boost/spirit/include/phoenix_object.hpp>
-#include <boost/spirit/include/phoenix_operator.hpp>
-#include <boost/spirit/include/phoenix_stl.hpp>
-
-#include <boost/lexical_cast.hpp>
-#include <boost/fusion/include/adapt_struct.hpp>
-#include <boost/fusion/include/std_pair.hpp>
-#include <boost/config/warning_disable.hpp>
-#include <boost/spirit/include/qi_numeric.hpp>
 #include <stan/lang/ast.hpp>
 #include <stan/lang/grammars/var_decls_grammar.hpp>
 #include <stan/lang/grammars/common_adaptors_def.hpp>
-
+#include <boost/spirit/include/qi.hpp>
+#include <boost/spirit/include/phoenix_core.hpp>
 #include <set>
 #include <string>
 #include <vector>
@@ -97,172 +84,9 @@ namespace stan {
 
   namespace lang {
 
-    struct validate_no_constraints_vis : public boost::static_visitor<bool> {
-      std::stringstream& error_msgs_;
-      explicit validate_no_constraints_vis(std::stringstream& error_msgs)
-        : error_msgs_(error_msgs) {
-      }
-      bool operator()(const nil& /*x*/) const {
-        error_msgs_ << "nil declarations not allowed";
-        return false;  // fail if arises
-      }
-      bool operator()(const int_var_decl& x) const {
-        if (x.range_.has_low() || x.range_.has_high()) {
-          error_msgs_ << "require unconstrained."
-                      << " found range constraint." << std::endl;
-          return false;
-        }
-        return true;
-      }
-      bool operator()(const double_var_decl& x) const {
-        if (x.range_.has_low() || x.range_.has_high()) {
-          error_msgs_ << "require unconstrained."
-                      << " found range constraint." << std::endl;
-          return false;
-        }
-        return true;
-      }
-      bool operator()(const vector_var_decl& x) const {
-        if (x.range_.has_low() || x.range_.has_high()) {
-          error_msgs_ << "require unconstrained."
-                      << " found range constraint." << std::endl;
-          return false;
-        }
-        return true;
-      }
-      bool operator()(const row_vector_var_decl& x) const {
-        if (x.range_.has_low() || x.range_.has_high()) {
-          error_msgs_ << "require unconstrained."
-                      << " found range constraint." << std::endl;
-          return false;
-        }
-        return true;
-      }
-      bool operator()(const matrix_var_decl& x) const {
-        if (x.range_.has_low() || x.range_.has_high()) {
-          error_msgs_ << "require unconstrained."
-                      << " found range constraint." << std::endl;
-          return false;
-        }
-        return true;
-      }
-      bool operator()(const unit_vector_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found unit_vector." << std::endl;
-        return false;
-      }
-      bool operator()(const simplex_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found simplex." << std::endl;
-        return false;
-      }
-      bool operator()(const ordered_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found ordered." << std::endl;
-        return false;
-      }
-      bool operator()(const positive_ordered_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found positive_ordered." << std::endl;
-        return false;
-      }
-      bool operator()(const cholesky_factor_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found cholesky_factor." << std::endl;
-        return false;
-      }
-      bool operator()(const cholesky_corr_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found cholesky_factor_corr." << std::endl;
-        return false;
-      }
-      bool operator()(const cov_matrix_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found cov_matrix." << std::endl;
-        return false;
-      }
-      bool operator()(const corr_matrix_var_decl& /*x*/) const {
-        error_msgs_ << "require unconstrained variable declaration."
-                    << " found corr_matrix." << std::endl;
-        return false;
-      }
-    };
 
-    struct data_only_expression : public boost::static_visitor<bool> {
-      std::stringstream& error_msgs_;
-      variable_map& var_map_;
-      data_only_expression(std::stringstream& error_msgs,
-                           variable_map& var_map)
-        : error_msgs_(error_msgs),
-          var_map_(var_map) {
-      }
-      bool operator()(const nil& /*e*/) const {
-        return true;
-      }
-      bool operator()(const int_literal& /*x*/) const {
-        return true;
-      }
-      bool operator()(const double_literal& /*x*/) const {
-        return true;
-      }
-      bool operator()(const array_literal& x) const {
-        for (size_t i = 0; i < x.args_.size(); ++i)
-          if (!boost::apply_visitor(*this, x.args_[i].expr_))
-            return false;
-        return true;
-      }
-      bool operator()(const variable& x) const {
-        var_origin origin = var_map_.get_origin(x.name_);
-        bool is_data = (origin == data_origin)
-          || (origin == transformed_data_origin)
-          || (origin == local_origin);
-        if (!is_data) {
-          error_msgs_ << "non-data variables not allowed"
-                      << " in dimension declarations."
-                      << std::endl
-                      << "     found variable=" << x.name_
-                      << "; declared in block=";
-          print_var_origin(error_msgs_, origin);
-          error_msgs_ << std::endl;
-        }
-        return is_data;
-      }
-      bool operator()(const integrate_ode& x) const {
-        return boost::apply_visitor(*this, x.y0_.expr_)
-          && boost::apply_visitor(*this, x.theta_.expr_);
-      }
-      bool operator()(const integrate_ode_cvode& x) const {
-        return boost::apply_visitor(*this, x.y0_.expr_)
-          && boost::apply_visitor(*this, x.theta_.expr_);
-      }
-      bool operator()(const fun& x) const {
-        for (size_t i = 0; i < x.args_.size(); ++i)
-          if (!boost::apply_visitor(*this, x.args_[i].expr_))
-            return false;
-        return true;
-      }
-      bool operator()(const index_op& x) const {
-        if (!boost::apply_visitor(*this, x.expr_.expr_))
-          return false;
-        for (size_t i = 0; i < x.dimss_.size(); ++i)
-          for (size_t j = 0; j < x.dimss_[i].size(); ++j)
-            if (!boost::apply_visitor(*this, x.dimss_[i][j].expr_))
-              return false;
-        return true;
-      }
-      bool operator()(const index_op_sliced& x) const {
-        return boost::apply_visitor(*this, x.expr_.expr_);
-      }
-      bool operator()(const binary_op& x) const {
-        return boost::apply_visitor(*this, x.left.expr_)
-          && boost::apply_visitor(*this, x.right.expr_);
-      }
-      bool operator()(const unary_op& x) const {
-        return boost::apply_visitor(*this, x.subject.expr_);
-      }
-    };
-
-
+    // TODO(carpenter): refactor this out of here; failed so far
+    //                  because of templating on R and T
     struct add_var {
       template <class> struct result;
       template <typename F, typename T1, typename T2, typename T3,
@@ -310,411 +134,13 @@ namespace stan {
     };
     boost::phoenix::function<add_var> add_var_f;
 
-
-    struct validate_decl_constraints {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3,
-                typename T4, typename T5>
-      struct result<F(T1, T2, T3, T4, T5)> { typedef void type; };
-
-      void operator()(const bool& allow_constraints,
-                      const bool& declaration_ok, const var_decl& var_decl,
-                      bool& pass, std::stringstream& error_msgs) const {
-        if (!declaration_ok) {
-          error_msgs << "Problem with declaration." << std::endl;
-          pass = false;
-          return;  // short-circuits test of constraints
-        }
-        if (allow_constraints) {
-          pass = true;
-          return;
-        }
-        validate_no_constraints_vis vis(error_msgs);
-        pass = boost::apply_visitor(vis, var_decl.decl_);
-      }
-    };
-    boost::phoenix::function<validate_decl_constraints>
-    validate_decl_constraints_f;
-
-    struct validate_identifier {
-      std::set<std::string> reserved_word_set_;
-      std::set<std::string> const_fun_name_set_;
-
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3>
-      struct result<F(T1, T2, T3)> { typedef void type; };
-
-      void reserve(const std::string& w) {
-        reserved_word_set_.insert(w);
-      }
-
-      template <typename S, typename T>
-      static bool contains(const S& s,
-                           const T& x) {
-        return s.find(x) != s.end();
-      }
-
-      bool identifier_exists(const std::string& identifier) const {
-        return contains(reserved_word_set_, identifier)
-          || (contains(function_signatures::instance().key_set(), identifier)
-              && !contains(const_fun_name_set_, identifier));
-      }
-
-      validate_identifier() {
-        // Constant functions which can be used as identifiers
-        const_fun_name_set_.insert("pi");
-        const_fun_name_set_.insert("e");
-        const_fun_name_set_.insert("sqrt2");
-        const_fun_name_set_.insert("log2");
-        const_fun_name_set_.insert("log10");
-        const_fun_name_set_.insert("not_a_number");
-        const_fun_name_set_.insert("positive_infinity");
-        const_fun_name_set_.insert("negative_infinity");
-        const_fun_name_set_.insert("epsilon");
-        const_fun_name_set_.insert("negative_epsilon");
-
-        // illegal identifiers
-        reserve("for");
-        reserve("in");
-        reserve("while");
-        reserve("repeat");
-        reserve("until");
-        reserve("if");
-        reserve("then");
-        reserve("else");
-        reserve("true");
-        reserve("false");
-
-        reserve("int");
-        reserve("real");
-        reserve("vector");
-        reserve("unit_vector");
-        reserve("simplex");
-        reserve("ordered");
-        reserve("positive_ordered");
-        reserve("row_vector");
-        reserve("matrix");
-        reserve("cholesky_factor_cov");
-        reserve("cholesky_factor_corr");
-        reserve("cov_matrix");
-        reserve("corr_matrix");
-
-
-        reserve("model");
-        reserve("data");
-        reserve("parameters");
-        reserve("quantities");
-        reserve("transformed");
-        reserve("generated");
-
-        reserve("var");
-        reserve("fvar");
-        reserve("STAN_MAJOR");
-        reserve("STAN_MINOR");
-        reserve("STAN_PATCH");
-        reserve("STAN_MATH_MAJOR");
-        reserve("STAN_MATH_MINOR");
-        reserve("STAN_MATH_PATCH");
-
-        reserve("alignas");
-        reserve("alignof");
-        reserve("and");
-        reserve("and_eq");
-        reserve("asm");
-        reserve("auto");
-        reserve("bitand");
-        reserve("bitor");
-        reserve("bool");
-        reserve("break");
-        reserve("case");
-        reserve("catch");
-        reserve("char");
-        reserve("char16_t");
-        reserve("char32_t");
-        reserve("class");
-        reserve("compl");
-        reserve("const");
-        reserve("constexpr");
-        reserve("const_cast");
-        reserve("continue");
-        reserve("decltype");
-        reserve("default");
-        reserve("delete");
-        reserve("do");
-        reserve("double");
-        reserve("dynamic_cast");
-        reserve("else");
-        reserve("enum");
-        reserve("explicit");
-        reserve("export");
-        reserve("extern");
-        reserve("false");
-        reserve("float");
-        reserve("for");
-        reserve("friend");
-        reserve("goto");
-        reserve("if");
-        reserve("inline");
-        reserve("int");
-        reserve("long");
-        reserve("mutable");
-        reserve("namespace");
-        reserve("new");
-        reserve("noexcept");
-        reserve("not");
-        reserve("not_eq");
-        reserve("nullptr");
-        reserve("operator");
-        reserve("or");
-        reserve("or_eq");
-        reserve("private");
-        reserve("protected");
-        reserve("public");
-        reserve("register");
-        reserve("reinterpret_cast");
-        reserve("return");
-        reserve("short");
-        reserve("signed");
-        reserve("sizeof");
-        reserve("static");
-        reserve("static_assert");
-        reserve("static_cast");
-        reserve("struct");
-        reserve("switch");
-        reserve("template");
-        reserve("this");
-        reserve("thread_local");
-        reserve("throw");
-        reserve("true");
-        reserve("try");
-        reserve("typedef");
-        reserve("typeid");
-        reserve("typename");
-        reserve("union");
-        reserve("unsigned");
-        reserve("using");
-        reserve("virtual");
-        reserve("void");
-        reserve("volatile");
-        reserve("wchar_t");
-        reserve("while");
-        reserve("xor");
-        reserve("xor_eq");
-
-        // function names declared in signatures
-        using stan::lang::function_signatures;
-        using std::set;
-        using std::string;
-        const function_signatures& sigs = function_signatures::instance();
-
-        set<string> fun_names = sigs.key_set();
-        for (set<string>::iterator it = fun_names.begin();
-             it != fun_names.end();
-             ++it)
-          if (!contains(const_fun_name_set_, *it))
-            reserve(*it);
-      }
-
-      void operator()(const std::string& identifier,
-                      bool& pass,
-                      std::stringstream& error_msgs) const {
-        int len = identifier.size();
-        if (len >= 2
-            && identifier[len-1] == '_'
-            && identifier[len-2] == '_') {
-          error_msgs << "variable identifier (name) may"
-                     << " not end in double underscore (__)"
-                     << std::endl
-                     << "    found identifer=" << identifier << std::endl;
-          pass = false;
-          return;
-        }
-        size_t period_position = identifier.find('.');
-        if (period_position != std::string::npos) {
-          error_msgs << "variable identifier may not contain a period (.)"
-                     << std::endl
-                     << "    found period at position (indexed from 0)="
-                     << period_position
-                     << std::endl
-                     << "    found identifier=" << identifier
-                     << std::endl;
-          pass = false;
-          return;
-        }
-        if (identifier_exists(identifier)) {
-          error_msgs << "variable identifier (name) may not be reserved word"
-                     << std::endl
-                     << "    found identifier=" << identifier
-                     << std::endl;
-          pass = false;
-          return;
-        }
-        pass = true;
-      }
-    };
-    boost::phoenix::function<validate_identifier> validate_identifier_f;
-
-    // copies single dimension from M to N if only M declared
-    struct copy_square_cholesky_dimension_if_necessary {
-      template <class> struct result;
-      template <typename F, typename T1>
-      struct result<F(T1)> { typedef void type; };
-      void operator()(cholesky_factor_var_decl& var_decl) const {
-        if (is_nil(var_decl.N_))
-          var_decl.N_ = var_decl.M_;
-      }
-    };
-    boost::phoenix::function<copy_square_cholesky_dimension_if_necessary>
-    copy_square_cholesky_dimension_if_necessary_f;
-
-    struct empty_range {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2>
-      struct result<F(T1, T2)> { typedef void type; };
-      void operator()(range& r, std::stringstream& /*error_msgs*/) const {
-        r = range();
-      }
-    };
-    boost::phoenix::function<empty_range> empty_range_f;
-
-    struct validate_int_expr {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3>
-      struct result<F(T1, T2, T3)> { typedef void type; };
-
-      void operator()(const expression& expr,
-                      bool& pass,
-                      std::stringstream& error_msgs) const {
-        if (!expr.expression_type().is_primitive_int()) {
-          error_msgs << "expression denoting integer required; found type="
-                     << expr.expression_type() << std::endl;
-
-          pass = false;
-          return;
-        }
-        pass = true;
-      }
-    };
-    boost::phoenix::function<validate_int_expr> validate_int_expr_f;
-
-    struct set_int_range_lower {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3, typename T4>
-      struct result<F(T1, T2, T3, T4)> { typedef void type; };
-      void operator()(range& range,
-                      const expression& expr,
-                      bool& pass,
-                      std::stringstream& error_msgs) const {
-        range.low_ = expr;
-        validate_int_expr validator;
-        validator(expr, pass, error_msgs);
-      }
-    };
-    boost::phoenix::function<set_int_range_lower> set_int_range_lower_f;
-
-    struct set_int_range_upper {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3, typename T4>
-      struct result<F(T1, T2, T3, T4)> { typedef void type; };
-      void operator()(range& range,
-                      const expression& expr,
-                      bool& pass,
-                      std::stringstream& error_msgs) const {
-        range.high_ = expr;
-        validate_int_expr validator;
-        validator(expr, pass, error_msgs);
-      }
-    };
-    boost::phoenix::function<set_int_range_upper> set_int_range_upper_f;
-
-    struct validate_int_data_expr {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3,
-                typename T4, typename T5>
-      struct result<F(T1, T2, T3, T4, T5)> { typedef void type; };
-
-      void operator()(const expression& expr,
-                      int var_origin,
-                      bool& pass,
-                      variable_map& var_map,
-                      std::stringstream& error_msgs) const {
-        if (!expr.expression_type().is_primitive_int()) {
-          error_msgs << "dimension declaration requires expression"
-                     << " denoting integer; found type="
-                     << expr.expression_type()
-                     << std::endl;
-          pass = false;
-        } else if (var_origin != local_origin) {
-          data_only_expression vis(error_msgs, var_map);
-          bool only_data_dimensions = boost::apply_visitor(vis, expr.expr_);
-          pass = only_data_dimensions;
-        } else {
-          // don't need to check data vs. parameter in dimensions for
-          // local variable declarations
-          pass = true;
-        }
-      }
-    };
-    boost::phoenix::function<validate_int_data_expr> validate_int_data_expr_f;
-
-    struct validate_double_expr {
-      template <typename T1, typename T2>
-      struct result { typedef bool type; };
-
-      bool operator()(const expression& expr,
-                      std::stringstream& error_msgs) const {
-        if (!expr.expression_type().is_primitive_double()
-            && !expr.expression_type().is_primitive_int()) {
-          error_msgs << "expression denoting real required; found type="
-                     << expr.expression_type() << std::endl;
-          return false;
-        }
-        return true;
-      }
-    };
-    boost::phoenix::function<validate_double_expr> validate_double_expr_f;
-
-
-    struct set_double_range_lower {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3, typename T4>
-      struct result<F(T1, T2, T3, T4)> { typedef void type; };
-      void operator()(range& range,
-                      const expression& expr,
-                      bool& pass,
-                      std::stringstream& error_msgs) const {
-        range.low_ = expr;
-        validate_double_expr validator;
-        pass = validator(expr, error_msgs);
-      }
-    };
-    boost::phoenix::function<set_double_range_lower> set_double_range_lower_f;
-
-    struct set_double_range_upper {
-      template <class> struct result;
-      template <typename F, typename T1, typename T2, typename T3, typename T4>
-      struct result<F(T1, T2, T3, T4)> { typedef void type; };
-      void operator()(range& range,
-                      const expression& expr,
-                      bool& pass,
-                      std::stringstream& error_msgs) const {
-        range.high_ = expr;
-        validate_double_expr validator;
-        pass = validator(expr, error_msgs);
-      }
-    };
-    boost::phoenix::function<set_double_range_upper> set_double_range_upper_f;
-
-
     template <typename Iterator>
     var_decls_grammar<Iterator>::var_decls_grammar(variable_map& var_map,
                                                std::stringstream& error_msgs)
       : var_decls_grammar::base_type(var_decls_r),
         var_map_(var_map),
         error_msgs_(error_msgs),
-        // expression_g allows full recursion
         expression_g(var_map, error_msgs),
-        // expression07_g disallows comparisons
         expression07_g(var_map, error_msgs, expression_g) {
       using boost::spirit::qi::_1;
       using boost::spirit::qi::_3;
